refactor to support 'network connection' interface, abstracting away connections to chargers
