+++ 00_Base/src/interfaces/messages/index.ts
@@ -36,7 +36,7 @@ export enum EventGroup {
 export { IMessage, Message } from "./Message";
 export { IMessageHandler } from "./MessageHandler";
 export { IMessageSender } from "./MessageSender";
-export { IMessageRouter } from "./MessageRouter";
+export { IMessageRouter, OcppMessageRouter } from "./MessageRouter";
 export { IMessageContext } from "./MessageContext";
 export { IMessageConfirmation } from "./MessageConfirmation";
 export { AbstractMessageHandler } from "./AbstractMessageHandler";
