+++ 00_Base/src/interfaces/modules/Module.ts
@@ -15,10 +15,8 @@ export interface IModule {
 
     sendCall(identifier: string, tenantId: string, action: CallAction, payload: OcppRequest, correlationId?: string, origin?: MessageOrigin): Promise<IMessageConfirmation>;
     sendCallResult(correlationId: string, identifier: string, tenantId: string, action: CallAction, payload: OcppResponse, origin?: MessageOrigin): Promise<IMessageConfirmation>;
-    sendCallResultWithMessage(message: IMessage<OcppRequest>, payload: OcppResponse): Promise<IMessageConfirmation>
     sendCallError(correlationId: string, identifier: string, tenantId: string, action: CallAction, error: OcppError, origin?: MessageOrigin): Promise<IMessageConfirmation>;
-    sendCallErrorWithMessage(message: IMessage<OcppRequest>, error: OcppError): Promise<IMessageConfirmation>;
-
+    
     handle(message: IMessage<OcppRequest | OcppResponse>, props?: HandlerProperties): Promise<void>;
     shutdown(): void;
