--- 02_Util/src/server/index.ts
@@ -1,7 +1,0 @@
-// Copyright (c) 2023 S44, LLC
-// Copyright Contributors to the CitrineOS Project
-//
-// SPDX-License-Identifier: Apache 2.0
-
-export { CentralSystemMessageHandler, OcppMessageRouter } from "./example/router";
-export { CentralSystemImpl } from "./example/server"
