+++ Server/src/index.ts
@@ -5,7 +5,7 @@
 
 import { ICache, ICentralSystem, IMessageHandler, IMessageSender, IModule, IModuleApi, SystemConfig } from '@citrineos/base';
 import { MonitoringModule, MonitoringModuleApi } from '@citrineos/monitoring';
-import { CentralSystemImpl, MemoryCache, RabbitMqReceiver, RabbitMqSender, initSwagger } from '@citrineos/util';
+import { Authenticator, CentralSystemImpl, MemoryCache, RabbitMqReceiver, RabbitMqSender, WebsocketNetworkConnection, initSwagger } from '@citrineos/util';
 import { JsonSchemaToTsProvider } from '@fastify/type-provider-json-schema-to-ts';
 import Ajv from "ajv";
 import addFormats from "ajv-formats"
@@ -19,6 +19,8 @@ import { EVDriverModule, EVDriverModuleApi } from '@citrineos/evdriver';
 import { ReportingModule, ReportingModuleApi } from '@citrineos/reporting';
 import { SmartChargingModule, SmartChargingModuleApi } from '@citrineos/smartcharging';
 import { sequelize } from '@citrineos/data';
+import { IAuthenticator } from '@citrineos/util/lib/centralsystem/authenticator/Authenticator';
+import { INetworkConnection } from '@citrineos/util/lib/centralsystem/networkconnection/WebsocketNetworkConnection';
 
 class CitrineOSServer {
 
@@ -28,6 +30,8 @@ class CitrineOSServer {
     private _config: SystemConfig;
     private _modules: Array<IModule>;
     private _apis: Array<IModuleApi>;
+    private _authenticator: IAuthenticator;
+    private _networkConnection: INetworkConnection;
     private _centralSystem: ICentralSystem;
     private _logger: Logger<ILogObj>;
     private _server: FastifyInstance;
@@ -84,7 +88,11 @@ class CitrineOSServer {
         // Force sync database
         sequelize.DefaultSequelizeInstance.getInstance(this._config, this._logger, true);
 
-        this._centralSystem = new CentralSystemImpl(this._config, this._cache, undefined, undefined, this._logger, ajv);
+        this._authenticator = new Authenticator(this._cache, this._logger, new sequelize.DeviceModelRepository(config, this._logger));
+
+        this._networkConnection = new WebsocketNetworkConnection(this._config.util.networkConnection.websocketServers, this._cache, this._logger, this._authenticator);
+
+        this._centralSystem = new CentralSystemImpl(this._config, this._cache, this._createSender(), this._createHandler(), this._networkConnection, this._logger, ajv);
 
         // Initialize modules & APIs
         // Always initialize APIs after SwaggerUI
@@ -108,7 +116,7 @@ class CitrineOSServer {
             new TransactionsModuleApi(transactionsModule, this._server, this._logger),
         ];
         if (this._config.modules.certificates) {
-           const certificatesModule = new CertificatesModule(this._config, this._cache, this._createSender(), this._createHandler(), this._logger)
+            const certificatesModule = new CertificatesModule(this._config, this._cache, this._createSender(), this._createHandler(), this._logger)
             this._modules.push(certificatesModule);
             this._apis.push(new CertificatesModuleApi(certificatesModule, this._server, this._logger));
         }
